backend/requirements.txt
# This file is used to specify the dependencies for the FastAPI application.
# It is recommended to pin versions to ensure compatibility.
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
python-dotenv==1.0.1
alembic==1.13.1